requirements.txt
loguru==0.7.0
fastapi==0.99.1
python-multipart==0.0.6
bcrypt==3.2.0
uvicorn==0.22.0
redis==4.6.0
pycryptodome==3.10.4
PyJWT==2.0.1
python-dotenv==1.0.0